The performance update.
No more Timer checking continuously the state of the server and
refreshing the output log.
This was really hard. Almost everything has been recode to do this.
More speed and lower memory usage thanks to the use of delegates for
refreshing the output and AppendText method.

Other changes:
Better Server Properties GUI.
Server auto-starts when opened with a shortcut.
"More" button menu for auto-input commands, including Reload and
Weather.
More options in Settings.