Officially "finished" developing
New option to read easly any of the compressed logs in "logs" folder.
New option to regenerate easly any world.
New option to replace for other one the CraftBukkit build that the
server has.
New setting to hide Max Memory box from the menu bar.
New colored dot in the statusbar representing the server status.

Warning if missing server.properties, so the user can configure a it
before server starts.
Resolved a bug which was preventing to download CraftBukkit in certain
conditions.
Resolved a bug which was naming incorrectly the shortcuts (duplicated
extension).

No more big functions are really needed. Now it's time of start to learn
C# and WPF.