Another update
Now the application is smart enough to find the craftbukkit file with
only the path. So you can select the server by file or by folder. That
was hard. This is because for not having to re-save favorites and
shortcuts everytime the server changes its craftbukkit build (so also
filename).

Added menu to open the server folder.
Press esc key to clear input textbox.