No need of a bukkit file on every server.
New option in settings for when select a server by folder, by favorite
or with a shortcut (the three methods that not specifies a Jar file,
only the folder), the CraftBukkit file that will be used for run the
server will be the set as default, located in
Documents\CBLauncher\versions.

Removed more commands menu, because all of them needs to specify a
world.

Other minor improvements in the code.